+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJAXYDB3R274MGY3XBH4ZY9M"
+        "01KJCGPRNJC4KSXY4K8A3EMHZ5"
     )
     r"""The unique identifier of the data source"""