+++ packages/orq-rc/RELEASES.md
@@ -6149,3 +6149,13 @@ Based on:
 - [python v4.5.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.9 - packages/orq-rc
+
+## 2026-02-26 08:25:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.730.1 (2.844.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.10] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.10 - packages/orq-rc